Add index.php, process_data.php, and recent_data.php to SQLorAPI namespace; implement super_function for unified API and SQL data retrieval
